+++ doc/whats-new.rst
@@ -32,6 +32,14 @@ Bug fixes
   coordinates so that they are centered in each pixel.
   By `Greg Brener <https://github.com/gbrener>`_.
 
+- :py:meth:`~xarray.Dataset.rename` method now doesn't throw errors
+  if some ``Variable`` is renamed to the same name as another ``Variable``
+  as long as that other ``Variable`` is also renamed. This method now
+  does throw when two ``Variables`` would end up with the same name
+  after the rename (since one of them would get overwritten in this
+  case). See (:issue:`1477`) for details.
+  By `Prakhar Goel <https://github.com/newt0311>`_.
+
 .. _whats-new.0.9.6:
 
 v0.9.6 (8 June 2017)
@@ -214,7 +222,7 @@ Enhancements
 
 - New aggregation on rolling objects :py:meth:`DataArray.rolling(...).count()`
   which providing a rolling count of valid values (:issue:`1138`).
-  
+
 Bug fixes
 ~~~~~~~~~
 - Rolling operations now keep preserve original dimension order (:issue:`1125`).